+++ lib/settings/customizer/header-footer.php
@@ -76,6 +76,31 @@ function mai_register_customizer_header_footer_settings( $wp_customize ) {
 		)
 	);
 
+	// Mobile menu width.
+	$wp_customize->add_setting(
+		_mai_customizer_get_field_name( $settings_field, 'mobile_menu_width' ),
+		array(
+			'default'           => absint( mai_get_default_option( 'mobile_menu_width' ) ),
+			'type'              => 'option',
+			'sanitize_callback' => 'absint',
+		)
+	);
+	$wp_customize->add_control( new Mai_Customize_Control_Slider( $wp_customize,
+		'mobile_menu_width',
+		array(
+			'label'       => esc_attr__( 'Mobile menu width', 'mai-theme-engine' ),
+			'description' => __( 'The largest screen/browser width at which the mobile menu becomes active.', 'mai-theme-engine' ),
+			'section'     => $section,
+			'settings'    => _mai_customizer_get_field_name( $settings_field, 'mobile_menu_width' ),
+			'priority'    => 10,
+			'input_attrs' => array(
+				'min'  => 0,    // Required.
+				'max'  => 1600, // Required.
+				'step' => 1,    // Required.
+			),
+		)
+	) );
+
 	// Footer widgets.
 	$wp_customize->add_setting(
 		_mai_customizer_get_field_name( $settings_field, 'footer_widget_count' ),